chore: reorganized some file locations
