+++ apps/api/src/app.module.ts
@@ -5,12 +5,12 @@ import { SentryGlobalFilter, SentryModule } from "@sentry/nestjs/setup";
 import { AppController } from "./app.controller";
 import { AssociationsModule } from "./associations/associations.module";
 import { AuthModule } from "./auth/auth.module";
+import { EntityNotFoundFilter } from "./common/filters/entity-not-found.filter";
+import { QueryFailedErrorFilter } from "./common/filters/query-failed-error.filter";
 import { CoursesModule } from "./courses/courses.module";
 import { DatabaseModule } from "./database/database.module";
 import { EventsModule } from "./events/events.module";
 import { FacultiesModule } from "./faculties/faculties.module";
-import { EntityNotFoundFilter } from "./filters/entity-not-found.filter";
-import { QueryFailedErrorFilter } from "./filters/query-failed-error.filter";
 import { RequestsModule } from "./requests/requests.module";
 import { ServicesModule } from "./services/services.module";
 import { UsersModule } from "./users/users.module";
