+++ apps/api/src/common/filters/query-failed-error.filter.spec.ts
