+++ apps/api/src/main.ts
@@ -4,7 +4,7 @@ import { ValidationPipe } from "@nestjs/common";
 import { NestFactory } from "@nestjs/core";
 import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
 import { AppModule } from "./app.module";
-import { EntityNotFoundFilter } from "./filters/entity-not-found.filter";
+import { EntityNotFoundFilter } from "./common/filters/entity-not-found.filter";
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
